Add psm event loop

+++ subprojects/LibPSM/Sce/Pss/Core/Timing/DeltaTime.hpp
@@ -1,17 +1,19 @@
 #ifndef LIB_PSS_DELTATIME_H
 #define LIB_PSS_DELTATIME_H 1
 
+#include <cstdint>
+
 namespace Sce::Pss::Core::Timing {
 	class DeltaTime{
 	private:
-		double minDelta;
-		double lastTime;
+		uint64_t minDelta;
+		uint64_t lastTime;
 	public:
 		DeltaTime(int minDelta);
 		~DeltaTime();
 
-		double MinDelta();
-		double CalculateDelta();
+		uint64_t MinDelta();
+		uint64_t CalculateDelta();
 	};
 }
